refactor: fix React fast refresh warnings
- Extracted useFormField hook to separate file
- Moved context declarations out of component files
- Separated SyncContext into context definition and provider
- Ensures proper hot module replacement during development

No functional changes, only code organization improvements
